requirements.txt
pytorch == 1.9.0
numpy == 1.20.2
scipy == 1.6.2
pandas == 1.2.4
transformers == 4.10.0
dotenv
tqdm
